Added search function.

+++ Day 29 - Building a Password Manager GUI App with Tkinter/main.py
@@ -21,7 +21,7 @@ def password_generator():
     password_entry.insert(0,password)
 # ---------------------------- SAVE PASSWORD ------------------------------- #
 def save_password():
-    website = website_entry.get()
+    website = website_entry.get().lower()
     username = username_entry.get()
     password = password_entry.get()
 
@@ -46,12 +46,30 @@ def save_password():
                         json.dump(new_data, pwd, indent=4)
                 
                 else:
-                    data.update(new_data)
-                    with open(r"Day 29 - Building a Password Manager GUI App with Tkinter\passwords.json","w") as pwd:
-                        json.dump(data, pwd, indent=4)
+                    if website_entry.get().lower() in data:
+                         warning_response = messagebox.askokcancel(title="Warning",message="Entry already exist; said entry will instead be replaced with the new one. continue?")
+                         if warning_response:
+                            data.update(new_data)
+                            with open(r"Day 29 - Building a Password Manager GUI App with Tkinter\passwords.json","w") as pwd:
+                                json.dump(data, pwd, indent=4)
                 finally:
                     website_entry.delete(0,END)
                     password_entry.delete(0,END)
+
+def find_website():
+    try:
+        with open (file=r"Day 29 - Building a Password Manager GUI App with Tkinter\passwords.json") as data:
+            data_base = json.load(data)
+    except FileNotFoundError:
+            messagebox.showerror(title="ERROR",message="Repository is empty. Please add NTLOGIN to the repository.")
+    else:
+        website = website_entry.get().lower()
+        if website in data_base:
+             data_result = data_base[website]
+             messagebox.showinfo(title=website.title(),message=f"Username: {data_result['username']}\nPassword: {data_result['password']}")
+        else:
+             messagebox.showerror(title="ERROR",message="Entry does not exist in the repository, consider saving one first.")
+
 
 # ---------------------------- UI SETUP ------------------------------- #
 window = Tk()
@@ -64,13 +82,13 @@ canvas.create_image(100,100,image=logo_img)
 canvas.grid(column=1,row=0)
 # ---------------------------- LABELS ------------------------------- #
 website_lb = Label(text="Website:")
-website_lb.grid(column=0,row=1)
+website_lb.grid(column=0,row=1,sticky="EW")
 
 username_lb = Label(text="Email/Username:")
-username_lb.grid(column=0,row=2)
+username_lb.grid(column=0,row=2,sticky="EW")
 
 password_lb = Label(text="Password")
-password_lb.grid(column=0,row=3)
+password_lb.grid(column=0,row=3,sticky="EW")
 # ---------------------------- ENTRIES ------------------------------- #
 website_entry = Entry(width=36)
 website_entry.grid(column=1,row=1,columnspan=2,sticky="EW")
@@ -83,8 +101,11 @@ username_entry.insert(0,"deepfreezer92@gmail.com")
 password_entry = Entry(width=21)
 password_entry.grid(column=1,row=3,sticky="EW")
 # ---------------------------- BUTTONS ------------------------------- #
+find_button = Button(text="Find Website", command=find_website)
+find_button.grid(column=2,row=1,sticky="EW")
+
 generate_button = Button(text="Generate Password",command=password_generator)
-generate_button.grid(column=2,row=3)
+generate_button.grid(column=2,row=3,sticky="EW")
 
 add_button = Button(text="Add",width=36,command=save_password)
 add_button.grid(column=1,row=4,columnspan=2,sticky="EW")
